Add a penalty to prevent the bot from playing direct damage spells to an opponent's face just because there's no other cards for it to play.
However, since hero abilities are treated like cards here (and we don't want a penalty on them), the TargetAbilitysDatabase (hero powers) had to be changed to index the hero powers by name rather than card id. This is fine since even though multiple hero powers can map to the same card name, it doesn't matter because we only care to know if it's a hero power or not.