Improve speed of Silverfish play actions under HearthRanger interface by queueing up multiple actions in the list returned to HearthRanger. Works in External mode and DLL mode.
The basic flow is:

1) Let Silverfish generate a list of actions as usual.
2) Instead of sending 1 action at a time to HearthRanger, we send as many as possible (in Bot::OnQueryBestMove)
3) We check "how many" actions we can send. Since it depends if the action has an expected outcome or a random one (in Ai::canQueueNextMoves), and other edge cases like minions spawning from other minions, etc.

Error handling:

There are times when HeartRanger will not execute all of our actions as we expect, either because we accidentally queued actions with unexpected outcomes (should no longer happen), or because of a bug/weirdness in the way HR does actions (example: any actions queued after playing buffs like Abusive Sergeant will never execute). When this happens our simulated board will no longer match the real board and we would have to re-simulate.

To prevent having to re-simulate we also have an error handling flow:

1) When sending list of actions to HR, we record that list internally (Bot::queuedMoveGuesses). We also record the number of actions sent.
2) We now listen for Bot::OnActionDone event to record the number of actions executed successfully.
3) When we encounter a board that does not match our expected board, we check the number of actions for a mismatch. If it is out of sync, we will restore the board from the saved list (Ai::restoreBestMoves)